refactor(app): make components smaller
