+++ src/app/[locale]/page.tsx
@@ -5,15 +5,10 @@ import { Grid, GridItem, Box } from "@chakra-ui/react";
 import { colors } from "../util/colors";
 import Header from "../components/Header";
 import ErDiagram from "../components/ErDiagram";
-import { ErrorMessage } from "../types/ErrorMessage";
-import ErrorTable from "../components/ErrorTable";
-import CodeEditor from "../components/CodeEditor";
+import ErrorReportingEditor from "../components/ErrorReportingEditor/ErrorReportingEditor";
 
 const Page = () => {
   const [erDoc, setErDoc] = useState<ER | null>(null);
-  const [semanticErrorMessages, setSemanticErrorMessages] = useState<
-    ErrorMessage[]
-  >([]);
 
   return (
     <Grid
@@ -56,28 +51,7 @@ const Page = () => {
           flexDir={"column"}
           overflow={"hidden"}
         >
-          <Box
-            resize="none"
-            pt={1}
-            flex={"1 1 auto"}
-            width="full"
-            height="max-content"
-            overflow="hidden"
-            bg={colors.textEditorBackground}
-          >
-            <CodeEditor
-              onErDocChange={setErDoc}
-              onSemanticErrorMessagesChange={setSemanticErrorMessages}
-            />
-          </Box>
-
-          <Box
-            height={"max-content"}
-            maxHeight={"30%"}
-            backgroundColor={colors.textEditorBackground}
-          >
-            <ErrorTable errors={semanticErrorMessages} />
-          </Box>
+          <ErrorReportingEditor onErDocChange={setErDoc} />
         </Box>
 
         <Box pt={1} width={"150%"} height="full">
