+++ src/app/components/ErrorReportingEditor/CodeEditor.tsx
@@ -1,12 +1,12 @@
 import Editor, { useMonaco, OnMount } from "@monaco-editor/react";
 import { languages, editor } from "monaco-types";
-import { MarkerSeverity } from "../types/CodeEditor";
+import { MarkerSeverity } from "../../types/CodeEditor";
 import { useTranslations } from "next-intl";
 import { Dispatch, useRef } from "react";
-import { getERDoc } from "../../ERDoc";
-import { ER } from "../../ERDoc/types/parser/ER";
-import { ErrorMessage } from "../types/ErrorMessage";
-import getErrorMessage from "../util/errorMessages";
+import { getERDoc } from "../../../ERDoc";
+import { ER } from "../../../ERDoc/types/parser/ER";
+import { ErrorMessage } from "../../types/ErrorMessage";
+import getErrorMessage from "../../util/errorMessages";
 import { Spinner } from "@chakra-ui/react";
 
 type EditorProps = {
@@ -37,7 +37,6 @@ const DEFAULT_THEME = "onedark";
 const editor_tokenizer: languages.IMonarchLanguage = {
   keywords: ["entity", "relation", "aggregation", "depends on", "extends"],
   keyKeywords: ["key", "pkey"],
-  keywordsRegex: /entity|relation|aggregation|depends on|extends/,
   ignoreCase: true,
   tokenizer: {
     root: [
